getPublicKeys: update for stax
