+++ src/getPublicKeys.h
@@ -30,4 +30,5 @@ typedef struct {
 
 handler_fn_t getPublicKeys_handleAPDU;
 
+void runGetOnePublicKeyUIFlow();
 #endif // H_CARDANO_APP_GET_PUBLIC_KEYS
